adds upvote component
