+++ src/components/upvote/upvote.component.tsx
@@ -1,34 +1,39 @@
+import { useState } from 'react';
 import { Heading, Pane, Text, Image } from 'evergreen-ui';
-import styled from 'styled-components';
-import UpvoteImage from '../../assets/icons/upvote.svg';
-export const UpVoteContainer = styled(Pane)`
-  height: 88px;
-  width: 88px;
-  padding: 9px 14px;
-  border: 2px solid #da552f;
-  border-radius: 5px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-`;
-
-export const Triangle = styled.div`
-  border-left: 10px solid transparent;
-  border-right: 10px solid transparent;
-  border-bottom: 10px solid #da552f;
-`;
+import UpvoteActive from '../../assets/icons/upvoteActive.svg';
+import UpvoteDefault from '../../assets/icons/upvoteDefault.svg';
+import { UpVoteContainer } from './upvote.component.styles';
 
 export const UpVote = () => {
+  const [count, setCount] = useState(false);
+  const [vote, setVote] = useState(0);
+
+  const handelCounter = () => {
+    setCount(true);
+    if (count) {
+      setVote(vote + 1);
+      setCount(false);
+    } else {
+      setVote(0);
+    }
+  };
+
   return (
-    <UpVoteContainer>
+    <UpVoteContainer count={count}>
       <Pane height={'72px'} width={'58px'}>
-        <Pane display='flex' flexDirection='column' alignItems='center' justifyContent='center'>
-          <Image src={UpvoteImage} />
+        <Pane
+          display='flex'
+          flexDirection='column'
+          alignItems='center'
+          justifyContent='center'
+          cursor='pointer'
+          onClick={handelCounter}
+        >
+          <Image src={!count ? UpvoteActive : UpvoteDefault} />
         </Pane>
         <Pane display='flex' flexDirection='column' alignItems='center' justifyContent='center' marginTop={'1.61px'}>
           <Heading size={500} padding={'2px'}>
-            10
+            {vote}
           </Heading>
           <Text size={300} marginTop={'1px'}>
             UPVOTES
